meaningful error on creating existing PO

+++ app/purchase/po_manager.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 import logging
 
-from flask import current_app
+from flask import current_app, jsonify
 from flask_security import current_user
 
 from app import db
@@ -12,6 +12,7 @@ from app.purchase.models import Company
 from app.purchase.models import PurchaseOrder, PurchaseOrderStatus
 from app.purchase.models.vendor_manager import PurchaseOrderVendorBase
 from app.purchase.signals import purchase_order_deleting, purchase_order_saving
+from exceptions import PurchaseOrderError
 
 def create_purchase_orders(order: Order, company: Company, address: Address,
                           vendor: PurchaseOrderVendorBase, contact_phone: str,
@@ -19,11 +20,15 @@ def create_purchase_orders(order: Order, company: Company, address: Address,
                           **kwargs
                          ) -> "list[PurchaseOrder]":
     logger = logging.getLogger('app.purchase.po_manager.create_purchase_order()')
-    if recreate_po:
-        for po in PurchaseOrder.query.filter(
-            PurchaseOrder.id.like(order.id.replace('ORD', 'PO') + '-%')):
-            purchase_order_deleting.send(po, **kwargs)
-            db.session.delete(po)
+    existing_pos = PurchaseOrder.query.filter(
+            PurchaseOrder.id.like(order.id.replace('ORD', 'PO') + '-%'))
+    if existing_pos.count() > 0:
+        if recreate_po:
+            for po in existing_pos:
+                purchase_order_deleting.send(po, **kwargs)
+                db.session.delete(po)
+        else:
+            raise PurchaseOrderError(message=f"Purchase orders for order {order.id} already exist")
     purchase_orders = []
     for suborder in order.suborders:
         purchase_order = PurchaseOrder(
